+++ polls/views/upsert_views.py
@@ -4,7 +4,7 @@ from django.shortcuts import get_object_or_404, redirect, render
 from django.urls import reverse
 from django.utils.translation import gettext as _
 from django.views import generic
-
+from cm_main.utils import check_edit_permission
 from ..models import Poll, Question
 from ..forms.upsert_forms import PollUpsertForm, QuestionUpsertForm
 
@@ -60,6 +60,7 @@ class PollUpdateView(LoginRequiredMixin, generic.UpdateView):
 
   def post(self, request, pk):
     poll = get_object_or_404(self.model, pk=pk)
+    check_edit_permission(request, poll.owner)
     # create a form instance from the request and save it
     form = self.form_class(request.POST, instance=poll)
     if form.is_valid():
@@ -75,6 +76,10 @@ class PollDeleteView(LoginRequiredMixin, generic.DeleteView):
   model = Poll
   success_url = "/polls/all/"
 
+  def post(self, request, pk):
+    check_edit_permission(request, self.get_object().owner)
+    return super().post(request, pk)
+
 
 class QuestionUpsertViewMixin(LoginRequiredMixin):
   model = Question
@@ -87,6 +92,7 @@ class QuestionCreateView(QuestionUpsertViewMixin, generic.CreateView):
     # create a form instance from the request and save it
     form = self.form_class(request.POST)
     poll = get_object_or_404(Poll, pk=poll_id)
+    check_edit_permission(request, poll.owner)
     if form.is_valid():
       form.instance.poll = poll
       form.save()
@@ -99,6 +105,7 @@ class QuestionUpdateView(QuestionUpsertViewMixin, generic.UpdateView):
   def post(self, request, poll_id, pk):
     # print("poll_id", poll_id, "pk", pk)
     question = get_object_or_404(self.model, pk=pk)
+    check_edit_permission(request, question.poll.owner)
     # print("Before save", question.__dict__)
     if question.poll.id != poll_id:
       raise ValueError('Question does not belong to this Poll')
@@ -117,6 +124,7 @@ class QuestionDeleteView(QuestionUpsertViewMixin, generic.DeleteView):
 
   def post(self, request, pk):
     question = get_object_or_404(self.model, pk=pk)
+    check_edit_permission(request, question.poll.owner)
     poll_id = question.poll.id
     question.delete()
     return redirect(reverse("polls:update_poll", args=(poll_id,)))
